+++ DbgHelpUtils/hex_dump.h
@@ -8,6 +8,6 @@ namespace dlg_help_utils
 
 namespace dlg_help_utils::hex_dump
 {
-    void hex_dump(std::wostream& os, void const* data, size_t length, size_t indent = 0, bool write_header = true, size_t bytes_per_row = 0x10, uint64_t offset = 0ULL);
-    void hex_dump(std::wostream& os, mini_dump_memory_stream& stream, size_t length, size_t indent = 0, bool write_header = true, size_t bytes_per_row = 0x10, uint64_t offset = 0ULL);
+    void hex_dump(std::wostream& os, void const* data, uint64_t length, size_t indent = 0, bool write_header = true, uint64_t bytes_per_row = 0x10, uint64_t offset = 0ULL);
+    void hex_dump(std::wostream& os, mini_dump_memory_stream& stream, uint64_t length, size_t indent = 0, bool write_header = true, uint64_t bytes_per_row = 0x10, uint64_t offset = 0ULL);
 }
